requirements.txt
numpy>=1.16
opencv-python>=4.1
Pillow>=6.1
pytesseract>=0.2
